Split up googlevr module
